requirements.txt
asgiref
Django
djangorestframework
pytz
sqlparse
psycopg2
drf-spectacular